backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client for API integrations
httpx==0.25.1

# Logging & Monitoring
python-json-logger==2.0.7

# PDF Generation (we'll add more as needed)
reportlab==4.0.4

# Markdown processing
markdown==3.5.1

# Date/Time handling
python-dateutil==2.8.2

# Environment variables
python-decouple==3.8